Correction on defect: #4, #7, #9

+++ SimpleTimeTracker/src/main/java/com/aknot/simpletimetracker/dialog/CategoryEditDialog.java
@@ -44,7 +44,7 @@ public class CategoryEditDialog extends Dialog {
 			public void onClick(final View view) {
 				currentCategory.setCategoryName(catNameField.getText().toString());
 				currentCategory.setDescription(catDescField.getText().toString());
-				currentCategory.setTargetHour(Integer.valueOf(catTargetField.getText().toString()));
+				currentCategory.setTargetHour(Double.valueOf(catTargetField.getText().toString()));
 				categoryActivity.onEditDialogSave(currentCategory);
 				dismiss();
 			}
